Consolidate search logic under Base Row
