+++ src/components/LogRow/ResmokeRow/index.tsx
@@ -1,9 +1,7 @@
-import { forwardRef, useMemo } from "react";
-import Highlight from "components/Highlight";
+import { forwardRef } from "react";
 import CollapsedRow from "components/LogRow//CollapsedRow";
 import BaseRow from "components/LogRow/BaseRow";
 import { isCollapsedRow } from "utils/collapsedRow";
-import renderHtml from "utils/renderHtml";
 import { BaseRowProps } from "../types";
 import { isLineInRange } from "../utils";
 
@@ -35,40 +33,16 @@ const ResmokeRow = forwardRef<any, BaseRowProps>((rowProps, ref) => {
       wrap={wrap}
       {...listRowProps}
       ref={ref}
+      data-cy-text="resmoke-row"
       highlightedLine={highlightedLine}
       lineNumber={line}
       scrollToLine={scrollToLine}
+      searchTerm={inRange ? searchTerm : undefined}
     >
-      <ProcessedResmokeRow
-        lineContent={lineContent}
-        searchTerm={inRange ? searchTerm : undefined}
-      />
+      {lineContent}
     </BaseRow>
   ) : null;
 });
-
-interface ProcessedResmokeRowProps {
-  lineContent: string;
-  searchTerm?: RegExp;
-}
-const ProcessedResmokeRow: React.FC<ProcessedResmokeRowProps> = ({
-  lineContent,
-  searchTerm,
-}) => {
-  const memoizedLogLine = useMemo(() => {
-    let render = lineContent;
-    if (searchTerm) {
-      render = render.replace(searchTerm, `<mark>$&</mark>`);
-    }
-    return renderHtml(render, {
-      transform: {
-        // @ts-expect-error - This is expecting a react component but its an Emotion component which are virtually the same thing
-        mark: Highlight,
-      },
-    });
-  }, [lineContent, searchTerm]);
-  return <span data-cy="resmoke-row">{memoizedLogLine}</span>;
-};
 
 ResmokeRow.displayName = "ResmokeRow";
